refactor(react): refactor react sdk to provide server side support
