+++ packages/react/src/index.ts
@@ -21,3 +21,4 @@ export * from './models/public-models';
 export {default as AsgardeoProvider} from './providers/AsgardeoProvider';
 export {default as useAuthentication} from './hooks/use-authentication';
 export {default as useOn} from './hooks/use-on';
+export {default as SPACryptoUtils} from './utils/crypto-utils';
